pyproject.toml

[project]
name = "fastapipg"
version = "0.1.0"
description = "fastapi,vue,blog"
authors = [
    {name = "jake Molly", email = "74493337+2412322029@users.noreply.github.com"},
]
dependencies = [
    "uvicorn~=0.22",
    "SQLAlchemy~=2.0",
    "slowapi~=0.1",
    "python-jose~=3.3",
    "PyMySQL~=1.0",
    "pydantic~=1.10",
    "psutil~=5.9",
    "Pillow~=9.5",
    "passlib~=1.7",
    "aiofiles~=23.1",
    "aiomysql~=0.1",
    "bcrypt>=4.0.1",
    "fastapi~=0.95",
    "limits~=3.5",
    "xlwt>=1.3.0",
    "python-multipart>=0.0.6",
    "websockets>=11.0.3",
    "ruamel-yaml>=0.17.28",
    "cryptography>=40.0.2",
]
requires-python = ">=3.11"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm.dev-dependencies]
test = [
    "bandit>=1.7.5",
]
